refactor: extract courses core into app/courses/ (highest-risk phase)
Ninth phase — the last and riskiest slice of the courses.py/main.py
decomposition. Course, AcademicYear, UserCourseOrder, and the
enrollments association table (the remainder of app/models/course.py
after kbm/content/discussion were already carved out) move to
app/courses/models.py; app/models/course.py is deleted entirely.
app/models/ now holds only school.py, user.py, issue.py, ticket.py,
token.py, activity_log.py — exactly the small set of genuinely
cross-cutting models the plan called out as not worth relocating.

Moved into app/courses/routes.py (two blueprints: courses_bp for /api
JSON routes — the CRUD/enroll/reorder logic that was already isolated
there — and courses_pages_bp, no prefix, for page renders):
- course_detail: the largest single view in the app (~110 lines),
  aggregating Quiz/Assignment/File/Link/Discussion into one "topics"
  feed with per-role, per-status filtering. Now imports models from
  app.quiz, app.assignment (via app.models), app.content, and
  app.discussion — normal in a Django-style layout where one app's view
  can reference another app's models.
- course_archives, course_import (page renders)
- api_get_courses, api_get_course_students, api_update_course_theme
  (main.py's /api routes)

Also finished sweeping main.py's remaining quiz/assignment lifecycle
routes into their owning apps (started in Phase 8, now complete):
quiz_detail/quiz_saved page renders and quiz archive/restore/delete
moved to a new quiz_pages_bp + quiz_bp in app/quiz/routes.py;
assignment archive/restore/delete moved into the existing
assignment_api_bp in app/assignment/routes.py. app/blueprints/main.py
is now just the true page remnant (index, dashboard, settings, history,
privacy-policy, sponsor, set-language) — ready for Phase 10.

Caught and fixed real bugs while doing this, invisible to `create_app()`
building successfully or pytest passing, since neither exercises these
exact render paths with populated data:
- course_detail's topics loop still called url_for('main.serve_file')
  and url_for('main.discussion_detail') — both endpoints moved to
  content_files/discussion_pages in Phases 7-8, so any course page with
  a file or a discussion would have thrown a Flask BuildError. Fixed to
  url_for('content_files.serve_file') / url_for('discussion_pages.discussion_detail').
- 4 quiz templates (quiz_detail.html, quiz_submission_detail.html x2,
  _question_form.html) still referenced url_for('main.serve_question_image')
  for question images — fixed to content_files.serve_question_image.
Found via a full repo-wide grep for every url_for('main....) and
url_for('courses....) reference (30+ hits) before starting this phase's
own endpoint renames, cross-checked against the actual routes now
registered — the exact discipline the plan's verification section
called for but that pytest alone can't provide, since these apps have
no template-rendering test coverage.

Verified: create_app() builds with all 222 routes intact (pure moves,
no route count change), SQLAlchemy metadata still has 33 tables, pytest
unchanged from baseline (25 pre-existing failures, 99 passing), and a
scripted smoke test (Flask test client, real seeded Quiz/Assignment/
File/Link/Discussion data, teacher AND student roles) exercises
course_detail, course_archives, course_import, quiz_detail (editor +
preview + student view), course_discussions, discussion_detail, and
the trash page end-to-end — all 200s, confirming every url_for() in the
aggregator view resolves correctly post-refactor.
