+++ app/courses/models.py
