+++ app/courses/templates/courses/course_archives.html
@@ -16,7 +16,7 @@
         <!-- Top Header -->
         <header class="bg-white dark:bg-gray-800 border-b border-gray-200 dark:border-gray-800 h-16 flex items-center justify-between px-8 sticky top-0 z-30">
             <div class="flex items-center space-x-4">
-                <a href="{{ url_for('main.course_detail', course_id=course.id) }}" class="p-2 text-gray-400 dark:text-gray-500 hover:text-primary-600 transition-colors">
+                <a href="{{ url_for('courses_pages.course_detail', course_id=course.id) }}" class="p-2 text-gray-400 dark:text-gray-500 hover:text-primary-600 transition-colors">
                     <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10 19l-7-7m0 0l7-7m-7 7h18"/>
                     </svg>
@@ -41,7 +41,7 @@
                             <h1 class="text-3xl font-black text-gray-900 dark:text-gray-100 mb-2">🗄️ Arsip</h1>
                             <p class="text-sm text-gray-500 dark:text-gray-400">Kuis, tugas, dan file yang diarsipkan tetap tersimpan dengan semua data.</p>
                         </div>
-                        <a href="{{ url_for('main.course_detail', course_id=course.id) }}" class="btn-duo btn-duo-ghost h-12 text-xs">
+                        <a href="{{ url_for('courses_pages.course_detail', course_id=course.id) }}" class="btn-duo btn-duo-ghost h-12 text-xs">
                             <svg class="w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                                 <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2.5" d="M10 19l-7-7m0 0l7-7m-7 7h18"/>
                             </svg>
@@ -304,7 +304,7 @@
                         </div>
                         <h3 class="text-xl font-bold text-gray-900 dark:text-gray-100 mb-2">Belum ada arsip</h3>
                         <p class="text-gray-500 dark:text-gray-400 mb-6 max-w-md mx-auto">Kuis, tugas, berkas, atau link yang diarsipkan akan muncul di sini.</p>
-                        <a href="{{ url_for('main.course_detail', course_id=course.id) }}" class="btn-duo btn-duo-blue h-14 text-sm inline-flex">
+                        <a href="{{ url_for('courses_pages.course_detail', course_id=course.id) }}" class="btn-duo btn-duo-blue h-14 text-sm inline-flex">
                             <svg class="w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                                 <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2.5" d="M10 19l-7-7m0 0l7-7m-7 7h18"/>
                             </svg>
