+++ app/discussion/templates/discussion/discussion_detail.html
@@ -9,7 +9,7 @@
         <div class="max-w-5xl mx-auto px-4 sm:px-6 lg:px-8">
             <div class="flex justify-between items-center h-16">
                 <div class="flex items-center">
-                    <a href="{{ url_for('main.course_detail', course_id=course.id) }}" class="flex items-center space-x-2 text-gray-600 hover:text-gray-900 transition-colors">
+                    <a href="{{ url_for('courses_pages.course_detail', course_id=course.id) }}" class="flex items-center space-x-2 text-gray-600 hover:text-gray-900 transition-colors">
                         <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                             <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15 19l-7-7 7-7"/>
                         </svg>
