+++ app/quiz/templates/quiz/quiz_detail.html
@@ -50,7 +50,7 @@
     <button onclick="confirmStart()" class="btn-duo btn-duo-primary w-full h-16 text-lg">
         MULAI KERJAKAN
     </button>
-    <a href="{{ url_for('main.course_detail', course_id=quiz.course_id) }}" class="block mt-6 text-sm font-bold text-gray-400 hover:text-gray-600">Kembali</a>
+    <a href="{{ url_for('courses_pages.course_detail', course_id=quiz.course_id) }}" class="block mt-6 text-sm font-bold text-gray-400 hover:text-gray-600">Kembali</a>
 </div>
 
 <!-- Main Quiz (Hidden initially) -->
@@ -61,7 +61,7 @@
             <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15 12a3 3 0 11-6 0 3 3 0 016 0z"/><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M2.458 12C3.732 7.943 7.523 5 12 5c4.478 0 8.268 2.943 9.542 7-1.274 4.057-5.064 7-9.542 7-4.477 0-8.268-2.943-9.542-7z"/></svg>
             <span class="text-sm font-medium">Ini adalah pratinjau. Jawaban Anda tidak akan disimpan.</span>
         </div>
-        <a href="{{ url_for('main.quiz_detail', quiz_id=quiz.id) }}" class="text-xs font-bold text-amber-900 bg-amber-200 hover:bg-amber-300 px-3 py-1 rounded-full transition-colors">Kembali ke Editor</a>
+        <a href="{{ url_for('quiz_pages.quiz_detail', quiz_id=quiz.id) }}" class="text-xs font-bold text-amber-900 bg-amber-200 hover:bg-amber-300 px-3 py-1 rounded-full transition-colors">Kembali ke Editor</a>
     </div>
     {% endif %}
 
@@ -111,7 +111,7 @@
                 
                 {% if question.image %}
                 <div class="mb-6 rounded-xl overflow-hidden border border-gray-100 shadow-sm max-w-2xl">
-                    <img src="{{ url_for('main.serve_question_image', course_id=question.quiz.course_id, filename=question.image) }}" 
+                    <img src="{{ url_for('content_files.serve_question_image', course_id=question.quiz.course_id, filename=question.image) }}" 
                          class="w-full h-auto object-cover" 
                          alt="Gambar Pertanyaan">
                 </div>
@@ -224,9 +224,9 @@
   window.questionsPerPage = {{ quiz.questions_per_page or 0 }};
   window.quizDescription = {{ (quiz.description or '') | tojson }};
   window.isPreview = {{ 'true' if is_preview else 'false' }};
-  window.quizExitUrl = {{ (url_for('main.quiz_detail', quiz_id=quiz.id) if is_preview else url_for('main.course_detail', course_id=quiz.course_id)) | tojson }};
-  window.quizEditorUrl = "{{ url_for('main.quiz_detail', quiz_id=quiz.id) }}";
-  window.quizReturnUrl = "{{ url_for('main.course_detail', course_id=quiz.course_id) }}";
+  window.quizExitUrl = {{ (url_for('quiz_pages.quiz_detail', quiz_id=quiz.id) if is_preview else url_for('courses_pages.course_detail', course_id=quiz.course_id)) | tojson }};
+  window.quizEditorUrl = "{{ url_for('quiz_pages.quiz_detail', quiz_id=quiz.id) }}";
+  window.quizReturnUrl = "{{ url_for('courses_pages.course_detail', course_id=quiz.course_id) }}";
   window.quizConfirmationMessage = {{ (quiz.confirmation_message or 'Jawaban Anda telah direkam.') | tojson }};
 </script>
 <script src="{{ url_for('static', filename='js/quiz_detail.js') }}?v={{ asset_hash('js/quiz_detail.js') }}"></script>
