+++ app/quiz/templates/quiz/quiz_submission_detail.html
@@ -10,7 +10,7 @@
 <div class="bg-white border-b border-gray-200 sticky top-0 z-50 px-6 py-4">
     <div class="flex items-center justify-between">
         <div class="flex items-center space-x-4">
-            <a href="{{ url_for('main.quiz_detail', quiz_id=submission.quiz_id) }}" class="text-gray-500 hover:bg-gray-100 p-2 rounded-full">
+            <a href="{{ url_for('quiz_pages.quiz_detail', quiz_id=submission.quiz_id) }}" class="text-gray-500 hover:bg-gray-100 p-2 rounded-full">
                 <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10 19l-7-7m0 0l7-7m-7 7h18"/></svg>
             </a>
             <div>
@@ -67,7 +67,7 @@
 
             {% if question.image %}
             <div class="mb-4 rounded-xl overflow-hidden border border-gray-100 max-w-md">
-                <img src="{{ url_for('main.serve_question_image', course_id=submission.quiz.course_id, filename=question.image) }}" class="w-full h-auto">
+                <img src="{{ url_for('content_files.serve_question_image', course_id=submission.quiz.course_id, filename=question.image) }}" class="w-full h-auto">
             </div>
             {% endif %}
 
@@ -131,7 +131,7 @@
 
                 {% elif question.question_type.name == 'UPLOAD' %}
                     {% if answer.answer_text %}
-                        <a href="{{ url_for('main.serve_question_image', course_id=submission.quiz.course_id, filename=answer.answer_text) }}" 
+                        <a href="{{ url_for('content_files.serve_question_image', course_id=submission.quiz.course_id, filename=answer.answer_text) }}" 
                            target="_blank" 
                            class="inline-flex items-center space-x-3 bg-white border border-gray-200 px-4 py-3 rounded-xl hover:bg-gray-50 transition-all group">
                             <div class="w-10 h-10 bg-blue-50 text-blue-600 rounded-lg flex items-center justify-center group-hover:bg-blue-600 group-hover:text-white transition-all">
