+++ app/templates/_sidebar.html
@@ -50,8 +50,8 @@
         <div class="pt-6">
             <p class="px-4 pb-2 text-[11px] font-black text-[#afafaf] uppercase tracking-[0.15em] sidebar-hide">Navigasi Kelas</p>
 
-            <a href="{{ url_for('main.course_detail', course_id=course.id) }}"
-               class="nav-link-duo {% if request.endpoint == 'main.course_detail' %}active{% endif %}">
+            <a href="{{ url_for('courses_pages.course_detail', course_id=course.id) }}"
+               class="nav-link-duo {% if request.endpoint == 'courses_pages.course_detail' %}active{% endif %}">
                 <div class="nav-icon-duo bg-[#58cc02]">
                     <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2.5" d="M12 6.253v13m0-13C10.832 5.477 9.246 5 7.5 5S4.168 5.477 3 6.253v13C4.168 18.477 5.754 18 7.5 18s3.332.477 4.5 1.253m0-13C13.168 5.477 14.754 5 16.5 5c1.747 0 3.332.477 4.5 1.253v13C19.832 18.477 18.247 18 16.5 18c-1.746 0-3.332.477-4.5 1.253"/>
@@ -73,7 +73,7 @@
 
             {% endif %}
 
-            <a href="{{ url_for('main.course_archives', course_id=course.id) }}"
+            <a href="{{ url_for('courses_pages.course_archives', course_id=course.id) }}"
                class="nav-link-duo {% if 'course_archives' in request.endpoint|default('') %}active{% endif %}">
                 <div class="nav-icon-duo bg-[#afafaf]">
                     <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
@@ -84,7 +84,7 @@
             </a>
 
             {% if current_user.role.value in ['guru', 'admin', 'super_admin'] and course.teacher_id == current_user.id %}
-            <a href="{{ url_for('main.course_import', course_id=course.id) }}"
+            <a href="{{ url_for('courses_pages.course_import', course_id=course.id) }}"
                class="nav-link-duo {% if request.endpoint == 'main.course_import' %}active{% endif %}">
                 <div class="nav-icon-duo bg-[#ce82ff]">
                     <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
