+++ app/trash/templates/trash/course_tps.html
@@ -10,7 +10,7 @@
     <div class="flex-1 flex flex-col min-w-0 overflow-hidden sidebar-transition">
         <header class="bg-white dark:bg-gray-900 border-b-2 border-slate-100 dark:border-gray-800 h-20 flex items-center justify-between px-8 sticky top-0 z-30">
             <div class="flex items-center gap-4 min-w-0">
-                <a href="{{ url_for('main.course_detail', course_id=course.id) }}"
+                <a href="{{ url_for('courses_pages.course_detail', course_id=course.id) }}"
                    class="p-2 text-[#afafaf] hover:text-[#4b4b4b] dark:hover:text-white">
                     <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2.5" d="M10 19l-7-7m0 0l7-7m-7 7h18"/></svg>
                 </a>
